spsegm(splm)	R Documentation

Spatial Simultaneous Equations

Description

Feasible generalized three stages least square estimator (FGS3SLS) of symultaneous systems of spatially interrelated cross sectional equations of the form:

Y = Y B + X C + bar{Y} L + U

U = bar{U} R + E

with Y = (y_1,...,y_m), X = (x_1,...,x_m), U = (u_1,...,u_m), bar{Y} = (bar{y}_1,...,bar{y}_m) and bar{y}_j = W y_j j=1,...,m, E = (e_1,...,e_m), bar{U} = (bar{u}_1,...,bar{u}_m) and bar{u}_j = W u_j j=1,...,m. B, C, L and R=diag_{j=1}^m (rho_j) are matrix of parameters

Usage


spsegm(formula,data=list(), spec="default", listw, method='spatialsim')

Arguments

formula 	 an object of class formula (see Details)

data 	 an object of class data.frame. An optional data frame containing the variables in the model

spec 	 a list used to define the specification (see Details)

listw 	 an object of class listw created for example by nb2listw

method 	 spatialsim

Details

The function can be specified with any number of equations. The number of equations is determined through the formula object. In particular, all the responses for all equations should be organized in a matrix and specified on the left hand side of the formula object. Analogously, all the explanatory variables from all equations should be specified on the right hand side of the formula. The number of equations is then retrieved from the dimension of the matrix containing the responses.

The argument spec should either be default or a list. When spec="default", all the equations in the system have the same explanatory variables (including the intercept). The intercept can be eliminated (for all the equations at the same time) as it is done in lm.

When spec=list it should have the same length as the number of equations (i.e. the number of columns of y in the formula object). The presence of the intercept is managed by spec (see Examples). When trying to eliminate the intercept in the classical way, if spec is a list an error message will occour.

Value

An object of class "splm".

coefficients 	 FG3SLS coefficients estimate of the model parameters (for all equations)

vcov 	 the variance covariance matrix of the estimated coefficients

type 	 'spsegm'

model 	 the matrix of the data used (responses in each equation are reported first, then the explanatory variables)

N 	 the number of cross-sectional observations

Eq 	 the number of equations in the system

k 	 the number of columns of the matrix of regressors (i.e. this corresponds to the number of explanatory variables in each equation only when spec="default")

call 	 the call used to create the object

terms 	 the terms object generated from formula and some optional arguments

Xnames 	 the names of the variables in the matrix of explanatory variables

Ynames 	 the names of the responses

spec 	 the argument spec

Author(s)

Gianfranco Pirasgpiras@mac.com

References

Kelejian, H.H. and Prucha, I.R. (2004) Estimation of Simultaneous systems of spatially interrelated cross sectional equations, Journal of Econometrics, 118, pages 27–50.

Kelejian, H.H. and Prucha, I.R. (1999) A Generalized Moments Estimator for the Autoregressive Parameter in a Spatial Model, International Economic Review, 40, pages 509–533.

Kelejian, H.H. and Prucha, I.R. (1999) A Generalized Spatial Two Stage Least Square Procedure for Estimating a Spatial Autoregressive Model with Autoregressive Disturbances, Journal of Real Estate Finance and Economics, 17, pages 99–121.

See Also

spregm

Examples


##set the dimension of the problem
n<-70
N<-n^2
ny<-3
nx<-3
##generate the spatial weights matrix
W<-nb2listw(cell2nb(n,n)) #requires spdep
##generate y and X, assign names to their columns and create a data.frame
y<-matrix(rnorm(N*ny),N,ny)
X<-matrix(runif(N*nx,-10,10),N,nx)
ynseq<-paste("y",seq(1:ny))
colnames(y)<-ynseq
xnseq<-paste("y",seq(1:nx))
colnames(X)<-xnseq      
data<-cbind(y,X)
data<-data.frame(data)
##spec='default'
SE<-spsegm( y~ X ,spec='default', listw=W)
summary(SE)
##spec=list(c(1,2,3),c(1,3,4),c(2,3,4))
##c(1,2,3): intercept (1) and first and second columns of X (2,3)
##c(1,3,4): intercept (1) and second and third columns of X (3,4)
##c(2,3,4): first, second and third columns of X (2,3,4)
SE<-spsegm(y~ X ,spec=list(c(1,2,3),c(1,3,4),c(2,3,4)), listw=W)

[Package splm version 0.1-1 Index]